+++ services/OneDrive.js
@@ -75,10 +75,14 @@ var SERVICE_CODE = {
         ["if==than", "$L10", "false", 2],
         ["create", "$L31", "Error", "Parent path does not exist.", "NotFound"],
         ["throwError", "$L31"],
-        ["if<=than", "$P3", 4000000, 2],
+        ["if<=than", "$P3", 4000000, 4],
         ["callFunc", "simpleUpload", "$P0", "$P1", "$P2", "$P3", "$P4"],
-        ["jumpRel", 1],
-        ["callFunc", "chunkedUpload", "$P0", "$P2", "$P1", "$P3", "$P4"]
+        ["if!=than", "$P5", null, 1],
+        ["callFunc", "updateMetadata", "$P0", "$P1", "$P5"],
+        ["jumpRel", 3],
+        ["callFunc", "chunkedUpload", "$P0", "$P2", "$P1", "$P3", "$P4"],
+        ["if!=than", "$P5", null, 1],
+        ["callFunc", "updateMetadata", "$P0", "$P1", "$P5"]
     ],
     "downloadSC": [
         ["callFunc", "validatePath", "$P0", "$P2"],
@@ -283,10 +287,14 @@ var SERVICE_CODE = {
         ["json.parse", "$L5", "$L4.responseBody"],
         ["set", "$L16", "$L5.value"],
         ["get", "$L15", "$L5", "@odata.nextLink"],
-        ["if!=than", "$L15", null, 14],
+        ["if!=than", "$L15", null, 18],
         ["create", "$L0", "Object"],
         ["set", "$L0.url", "$L15"],
         ["set", "$L0.method", "GET"],
+        ["create", "$L10", "Object"],
+        ["string.concat", "$L10.Authorization", "Bearer ", "$S0.accessToken"],
+        ["set", "$L10.Accept", "application/json"],
+        ["set", "$L0.requestHeaders", "$L10"],
         ["http.requestCall", "$L4", "$L0"],
         ["callFunc", "validateResponse", "$P0", "$L4"],
         ["json.parse", "$L5", "$L4.responseBody"],
@@ -297,7 +305,7 @@ var SERVICE_CODE = {
         ["push", "$L16", "$L8"],
         ["math.add", "$L7", "$L7", 1],
         ["jumpRel", -5],
-        ["jumpRel", -16],
+        ["jumpRel", -20],
         ["create", "$L6", "Array"],
         ["create", "$L7", "Number", 0],
         ["create", "$L8", "Number", 0],
@@ -380,6 +388,10 @@ var SERVICE_CODE = {
         ["create", "$L0", "Object"],
         ["set", "$L0.url", "$P0.paginationCache.cursor"],
         ["set", "$L0.method", "GET"],
+        ["create", "$L10", "Object"],
+        ["string.concat", "$L10.Authorization", "Bearer ", "$S0.accessToken"],
+        ["set", "$L10.Accept", "application/json"],
+        ["set", "$L0.requestHeaders", "$L10"],
         ["http.requestCall", "$L1", "$L0"],
         ["callFunc", "validateResponse", "$P0", "$L1"],
         ["json.parse", "$L2", "$L1.responseBody"],
@@ -388,7 +400,7 @@ var SERVICE_CODE = {
         ["create", "$P0.paginationCache.metaCache", "Array"],
         ["callFunc", "processRawMeta", "$P0", "$P0.paginationCache.metaCache", "$L2.value", "$P2"],
         ["get", "$P0.paginationCache.cursor", "$L2", "@odata.nextLink"],
-        ["jumpRel", -57]
+        ["jumpRel", -61]
     ],
     "getAllocation": [
         ["callFunc", "checkAuth", "$P0", "$L20"],
@@ -608,6 +620,12 @@ var SERVICE_CODE = {
         ["if!=than", "$L5.lastModifiedDateTime", null, 2],
         ["create", "$L10", "Date", "$L5.lastModifiedDateTime"],
         ["set", "$P1.modifiedAt", "$L10.time"],
+        ["if!=than", "$L5.fileSystemInfo", null, 5],
+        ["create", "$L14", "Object"],
+        ["set", "$L14", "$L5.fileSystemInfo"],
+        ["if!=than", "$L14", null, 2],
+        ["create", "$L15", "Date", "$L14.lastModifiedDateTime"],
+        ["set", "$P1.contentModifiedAt", "$L15.time"],
         ["if==than", "$L5.folder", null, 2],
         ["set", "$P1.Folder", 0],
         ["jumpRel", 1],
@@ -864,6 +882,48 @@ var SERVICE_CODE = {
         ["math.add", "$L2", "$L2", 1],
         ["jumpRel", -8],
         ["set", "$P1", "$L4"]
+    ],
+    "CloudStorage:uploadWithContentModifiedDate": [
+        ["callFunc", "checkNull", "$P0", "$P2", "$P5"],
+        ["create", "$L0", "Date"],
+        ["set", "$L0.time", "$P5"],
+        ["set", "$L1", "$L0.rfcTimeUsingFormat2"],
+        ["callFunc", "uploadSC", "$P0", "$P1", "$P2", "$P3", "$P4", "$L1"]
+    ],
+    "updateMetadata": [
+        ["create", "$L0", "Object"],
+        ["create", "$L1", "String"],
+        ["callFunc", "urlEncode", "$P0", "$L2", "$P1"],
+        ["string.concat", "$L1", "$P0.baseUrl", "/drive/root:", "$L2"],
+        ["set", "$L0.url", "$L1"],
+        ["set", "$L0.method", "GET"],
+        ["create", "$L10", "Object"],
+        ["string.concat", "$L10.Authorization", "Bearer ", "$S0.accessToken"],
+        ["set", "$L10.Accept", "application/json"],
+        ["set", "$L0.requestHeaders", "$L10"],
+        ["create", "$L4", "Object"],
+        ["http.requestCall", "$L4", "$L0"],
+        ["callFunc", "validateResponse", "$P0", "$L4"],
+        ["create", "$L5", "Object"],
+        ["json.parse", "$L5", "$L4.responseBody"],
+        ["get", "$L3", "$L5.id"],
+        ["create", "$L6", "Object"],
+        ["string.concat", "$L6.url", "$P0.baseUrl", "/drive/items/", "$L3"],
+        ["create", "$L7", "Object"],
+        ["create", "$L8", "Object"],
+        ["set", "$L8.lastModifiedDateTime", "$P2"],
+        ["set", "$L7.fileSystemInfo", "$L8"],
+        ["json.stringify", "$L7", "$L7"],
+        ["stream.stringToStream", "$L7", "$L7"],
+        ["create", "$L9", "Object"],
+        ["set", "$L9.Content-Type", "application/json"],
+        ["string.concat", "$L9.Authorization", "Bearer ", "$S0.accessToken"],
+        ["set", "$L6.method", "PATCH"],
+        ["set", "$L6.requestHeaders", "$L9"],
+        ["set", "$L6.requestBody", "$L7"],
+        ["create", "$L11", "Object"],
+        ["http.requestCall", "$L11", "$L6"],
+        ["callFunc", "validateResponse", "$P0", "$L11"]
     ]
 };
 var OneDrive = (function () {
@@ -1138,6 +1198,20 @@ var OneDrive = (function () {
         var ip = new Interpreter_1.Interpreter(new Sandbox_1.Sandbox(SERVICE_CODE, this.persistentStorage, this.instanceDependencyStorage));
         ip.callFunction("Authenticating:logout", this.interpreterStorage).then(function () {
             Helper_1.Helper.checkSandboxError(ip, "OneDrive", "logout");
+        }).then(function () {
+            var res;
+            if (callback != null && typeof callback === "function")
+                callback(undefined, res);
+        }, function (err) {
+            if (callback != null && typeof callback === "function")
+                callback(err);
+        });
+    };
+    OneDrive.prototype.uploadWithContentModifiedDate = function (filePath, stream, size, overwrite, contentModifiedDate, callback) {
+        Statistics_1.Statistics.addCall("OneDrive", "uploadWithContentModifiedDate");
+        var ip = new Interpreter_1.Interpreter(new Sandbox_1.Sandbox(SERVICE_CODE, this.persistentStorage, this.instanceDependencyStorage));
+        ip.callFunction("CloudStorage:uploadWithContentModifiedDate", this.interpreterStorage, filePath, stream, size, overwrite ? 1 : 0, contentModifiedDate).then(function () {
+            Helper_1.Helper.checkSandboxError(ip, "OneDrive", "uploadWithContentModifiedDate");
         }).then(function () {
             var res;
             if (callback != null && typeof callback === "function")